OC-126: split humongous InstrumentationTest into InstrumentationTestBase and over a dozen of test classes; add a test of complexityOfTernaryOperator in ExpressionComplexityCounterTest; remove counting ternaryComplexity=1 in conditionalExpression as it's already accounted for in the ExpressionInfo.fromTokens counter
