+++ clover-core/src/test/groovy/com/atlassian/clover/instr/java/ExpressionComplexityCounterTest.groovy
@@ -3,12 +3,16 @@ package com.atlassian.clover.instr.java
 import org.junit.Test
 
 import static com.atlassian.clover.instr.java.JavaTokenTypes.BOR
+import static com.atlassian.clover.instr.java.JavaTokenTypes.COLON
 import static com.atlassian.clover.instr.java.JavaTokenTypes.EQUAL
 import static com.atlassian.clover.instr.java.JavaTokenTypes.IDENT
 import static com.atlassian.clover.instr.java.JavaTokenTypes.INT_LITERAL
 import static com.atlassian.clover.instr.java.JavaTokenTypes.LAND
 import static com.atlassian.clover.instr.java.JavaTokenTypes.LOR
+import static com.atlassian.clover.instr.java.JavaTokenTypes.LT
 import static com.atlassian.clover.instr.java.JavaTokenTypes.LPAREN
+import static com.atlassian.clover.instr.java.JavaTokenTypes.MINUS
+import static com.atlassian.clover.instr.java.JavaTokenTypes.QUESTION
 import static com.atlassian.clover.instr.java.JavaTokenTypes.RPAREN
 import static org.junit.Assert.assertEquals
 
@@ -18,7 +22,7 @@ class ExpressionComplexityCounterTest {
     void complexityOfOne() {
         ExpressionComplexityCounter counter = new ExpressionComplexityCounter(1)
 
-        // "a == 5 | 12" - bit OR
+        // "a == 5 | 12" - bit OR is not a branch
         counter.accept(new CloverToken(IDENT, "a"))
         counter.accept(new CloverToken(EQUAL, "=="))
         counter.accept(new CloverToken(INT_LITERAL, "5"))
@@ -32,7 +36,7 @@ class ExpressionComplexityCounterTest {
     void complexityOfThree() {
         ExpressionComplexityCounter counter = new ExpressionComplexityCounter(1)
 
-        // "a || (b && c)" - bit OR
+        // "a || (b && c)" - logical AND / OR are branches
         counter.accept(new CloverToken(IDENT, "a"))
         counter.accept(new CloverToken(LOR, "||"))
         counter.accept(new CloverToken(LPAREN, "("))
@@ -42,6 +46,23 @@ class ExpressionComplexityCounterTest {
         counter.accept(new CloverToken(RPAREN, ")"))
 
         assertEquals(3, counter.getComplexity())
+    }
+
+    @Test
+    void complexityOfTernaryOperator() {
+        ExpressionComplexityCounter counter = new ExpressionComplexityCounter(1)
+
+        // "a < 0 ? -1 : 1" - ternary has two branches so one cycle
+        counter.accept(new CloverToken(IDENT, "a"))
+        counter.accept(new CloverToken(LT, "<"))
+        counter.accept(new CloverToken(INT_LITERAL, "0"))
+        counter.accept(new CloverToken(QUESTION, "?"))
+        counter.accept(new CloverToken(MINUS, "-"))
+        counter.accept(new CloverToken(INT_LITERAL, "1"))
+        counter.accept(new CloverToken(COLON, ":"))
+        counter.accept(new CloverToken(INT_LITERAL, "1"))
+
+        assertEquals(1, counter.getComplexity())
     }
 
 }
